feat: cli command group: collateral
