+++ tests/test_cli/test_transactions.py
@@ -8,17 +8,3 @@ def test_transaction_get(runner):
 
     result = runner.invoke(drv, ["transaction", "get", "514aea9b-c86b-49e5-ba89-620b38477cdd"])
     assert result.exit_code == 0, f"Command failed with output:\n{result.output}"
-
-
-def test_transaction_deposit_to_subaccount(runner):
-    """Test: `drv transaction deposit-to-subaccount`"""
-
-    result = runner.invoke(drv, ["transaction", "deposit-to-subaccount", "0.1", "USDC"])
-    assert result.exit_code == 0, f"Command failed with output:\n{result.output}"
-
-
-def test_transaction_withdraw_from_subaccount(runner):
-    """Test: `drv transaction withdraw-from-subaccount`"""
-
-    result = runner.invoke(drv, ["transaction", "withdraw-from-subaccount", "0.1", "USDC"])
-    assert result.exit_code == 0, f"Command failed with output:\n{result.output}"
